Reworked all the container, thanks zag, any issues please report

+++ init/60_set_config.sh
@@ -1,23 +1,7 @@
 #!/bin/bash
 
-if [ ! -f /config/config.conf ]; then
-  cp /defaults/config.conf /config/config.conf
-fi
+[[ ! -f /config/config.conf ]] && cp /defaults/config.conf /config/config.conf
+[[ ! -f /app/plexemail/web/index.html ]] && /sbin/setuser abc cp /defaults/index.html /app/plexemail/web/index.html
+[[ ! -f /etc/cron.d/plexemail ]] && cp /defaults/plexemail /etc/cron.d/plexemail
 
-if [ ! -f /app/plexemail/web/index.html ]; then
-	cp /config/www/index.html /app/plexemail/web/index.html
-fi
-
-if [ ! -f /config/testscripts/send_one_time_email.sh ]; then
-	mkdir -p /config/testscripts/
-  cp /defaults/send_one_time_email.sh /config/testscripts/send_one_time_email.sh
-fi
-
-if [ ! -f /config/testscripts/test_plexemail.sh ]; then
-  mkdir -p /config/testscripts/
-  cp /defaults/test_plexemail.sh /config/testscripts/test_plexemail.sh
-fi
-
-crontab /defaults/plexemail
-
-chown abc:abc -R /config
+chown -R abc:abc /config
